BUG FIX 872165: [XClient][NetCore]It throws 405 exception when invoking CreateIfNotExistsAsync against azure file root directory. (#92)
Invoking CreateIfNotExistsAsync on a root directory should not throw the error 405. If a share exists, the root directory always exists, so calling CreateIfNotExistsAsync on an existing root directory should return false, meaning nothing was created. If a share is nonexistent, then CreateIfNotExistsAsync should return an error 404. This behavior is consistent with the Desktop version of .NET library.