[.NET][8.1.2 Hotfix]Download hang/ SendStream leak
-- Client would hang when facing half-open connections during reading
partial response from server, in which case the maximum execution time
passed as a token to the ReadAsync operation to Network stream was
ignored(this issue will be tracked with .Net team separately). As a
workaround we are adding a control to cancel the wait on the read
operation (note that the read operation itself might still be going on but
the flow will not be stopped indefinitely due to it). Stephen Toub has
provided a tidy extension method in his blog post referring to a similar
issue, which we are using in the TaskExtensions class(the reference is
mentioned in the CR).

--A memory leak where the send stream was not disposed in case of
failures/retries and also for all table operations

--assembly version update to 8.1.2