Fix #928 Fix formatting of GetRootContainerReference xmldoc remarks
This change fixes the formatting of the `<remarks>` field of the xmldoc for CloudBlobClient.GetRootContainerReference. It appears fine when reading the source code, but on docs.microsoft.com, the second line appears indented. I think this may be because of the newline in this block.
This newline in this block has been removed. There are other LOC in this file that are around or above the character count per line, if keeping a consistent ruler is an issue.