Avoid unnecessary exception in TimeoutStream
Underlying stream implementation may throw if we try to set a read/write timeout on a non-readable or non-writable stream. Add an additional check to prevent this exception from being thrown.

Since this gets hit on every blob read, the impact is fairly significant for services which download many blobs.